requirements.txt
streamlit
google-generativeai
folium
streamlit-folium
pandas
python-dotenv
pydantic
